refactor: consolidate chart functionality
