+++ app/Http/Livewire/Home/Chart.php
@@ -7,9 +7,8 @@ namespace App\Http\Livewire\Home;
 use App\Enums\StatsPeriods;
 use App\Facades\Network;
 use App\Facades\Settings;
-use App\Http\Livewire\Concerns\AvailablePeriods;
 use App\Http\Livewire\Concerns\GetsCurrentPrice;
-use App\Http\Livewire\Concerns\StatisticsChart;
+use App\Http\Livewire\Concerns\HandlesChart;
 use App\Services\NumberFormatter as ServiceNumberFormatter;
 use Illuminate\Support\Collection;
 use Illuminate\View\View;
@@ -17,32 +16,15 @@ use Livewire\Component;
 
 final class Chart extends Component
 {
-    use AvailablePeriods;
+    use HandlesChart;
     use GetsCurrentPrice;
-    use StatisticsChart;
 
-    public bool $show = true;
-
-    public string $period = '';
-
-    public string $refreshInterval = '';
-
-    /** @var mixed */
-    protected $listeners = [
-        'currencyChanged' => '$refresh',
-        'themeChanged'    => '$refresh',
-        'updateChart'     => '$refresh',
-    ];
-
-    public function mount(): void
+    public function render(): View | string
     {
-        $this->refreshInterval = (string) config('arkscan.statistics.refreshInterval', '60');
-        $this->period          = $this->defaultPeriod();
-        $this->show            = Network::canBeExchanged();
-    }
+        if (! Network::canBeExchanged()) {
+            return '<div></div>';
+        }
 
-    public function render(): View
-    {
         $currency = Settings::currency();
 
         $currentPrice = $this->getPrice($currency);
@@ -72,15 +54,6 @@ final class Chart extends Component
         }
 
         $this->period = $period;
-    }
-
-    private function mainValueVariation(array $dataset): string
-    {
-        // Determine difference based on first datapoint
-        $initialValue = collect($dataset)->first();
-        $currentValue = $this->getPrice(Settings::currency());
-
-        return $initialValue > $currentValue ? 'red' : 'green';
     }
 
     private function availablePeriods(): array
