+++ app/Http/Livewire/Stats/InsightAllTimeFeesCollected.php
@@ -19,8 +19,6 @@ final class InsightAllTimeFeesCollected extends Component
     use ChartNumberFormatters;
     use StatisticsChart;
 
-    public string $period = '';
-
     public string $refreshInterval = '';
 
     public string $cache = '';
